app/bq-importer/requirements.txt
google-cloud-bigquery==1.24.0
google-cloud-storage==1.26.0
